chore: lint and deploy
