+++ .prettierignore
@@ -1,0 +1,4 @@
+dist/
+node_modules/
+*.lock
+LICENSE
