+++ tsconfig.json
@@ -11,6 +11,7 @@
         "module": "ESNext",
         "moduleResolution": "Node",
         "resolveJsonModule": true,
+        "strictNullChecks": true,
         "isolatedModules": false,
         "noEmit": true,
         "jsx": "react"
